+++ R/leiden.R
@@ -6,12 +6,13 @@ NULL
 ##'
 ##' @description Implements the Leiden clustering algorithm in R using reticulate to run the Python version. Requires the python "leidenalg" and "igraph" modules to be installed. Returns a vector of partition indices.
 ##' Windows users can still this with devtools::install_github("rstudio/reticulate", ref = "86ebb56"); reticulate::use_condaenv("r-reticulate"); reticulate::conda_install("r-reticulate", "leidenalg", channel = "vtraag")
-##' @param object An adjacency matrix compatible with \code{\link[igraph]{igraph}} object or an input graph as an \code{\link[igraph]{igraph}} object (e.g., shared nearest neighbours).
+##' @param object An adjacency matrix compatible with \code{\link[igraph]{igraph}} object or an input graph as an \code{\link[igraph]{igraph}} object (e.g., shared nearest neighbours). A list of multiple graph objects can be passed for multiplex community detection.
 ##' @param partition_type Type of partition to use. Defaults to RBConfigurationVertexPartition. Options include: ModularityVertexPartition, RBERVertexPartition, CPMVertexPartition, MutableVertexPartition, SignificanceVertexPartition, SurpriseVertexPartition, ModularityVertexPartition.Bipartite, CPMVertexPartition.Bipartite (see the Leiden python module documentation for more details)
 ##' @param initial_membership,weights,node_sizes Parameters to pass to the Python leidenalg function (defaults initial_membership=None, weights=None). Weights are derived from weighted igraph objects and non-zero integer values of adjacency matrices.
 ##' @param resolution_parameter A parameter controlling the coarseness of the clusters
 ##' @param seed Seed for the random number generator. By default uses a random seed if nothing is specified.
 ##' @param n_iterations Number of iterations to run the Leiden algorithm. By default, 2 iterations are run. If the number of iterations is negative, the Leiden algorithm is run until an iteration in which there was no improvement.
+##' @param max_comm_size (non-negative int) – Maximal total size of nodes in a community. If zero (the default), then communities can be of any size.
 ##' @param degree_as_node_size (defaults to FALSE). If True use degree as node size instead of 1, to mimic modularity for Bipartite graphs.
 ##' @param laplacian (defaults to FALSE). Derive edge weights from the Laplacian matrix.
 ##' @return A partition of clusters as a vector of integers
@@ -53,29 +54,29 @@ NULL
 ##' # generate (unweighted) igraph object in R
 ##' library("igraph")
 ##' adjacency_matrix[adjacency_matrix > 1] <- 1
-##' snn_graph <- graph_from_adjacency_matrix(adjacency_matrix)
-##' partition <- leiden(snn_graph)
+##' my_graph <- graph_from_adjacency_matrix(adjacency_matrix)
+##' partition <- leiden(my_graph)
 ##' table(partition)
 ##'
 ##' # generate (weighted) igraph object in R
 ##' library("igraph")
 ##' adjacency_matrix[adjacency_matrix >= 1] <- weights
-##' snn_graph <- graph_from_adjacency_matrix(adjacency_matrix, weighted = TRUE)
-##' partition <- leiden(snn_graph)
+##' my_graph <- graph_from_adjacency_matrix(adjacency_matrix, weighted = TRUE)
+##' partition <- leiden(my_graph)
 ##' table(partition)
 ##'
 ##' # pass weights to python leidenalg
 ##' adjacency_matrix[adjacency_matrix >= 1 ] <- 1
-##' snn_graph <- graph_from_adjacency_matrix(adjacency_matrix, weighted = NULL)
+##' my_graph <- graph_from_adjacency_matrix(adjacency_matrix, weighted = NULL)
 ##' weights <- sample(1:10, sum(adjacency_matrix!=0), replace=TRUE)
-##' partition <- leiden(snn_graph, weights = weights)
+##' partition <- leiden(my_graph, weights = weights)
 ##' table(partition)
 ##'
 ##' # run only if python is available (for testing)
 ##' }
 ##'
 ##' @keywords graph network igraph mvtnorm simulation
-##' @importFrom reticulate import py_to_r r_to_py
+##' @importFrom reticulate import py_to_r r_to_py py_has_attr py_get_attr py_set_attr
 ##' @rdname leiden
 ##' @export
 leiden <- function(object,
@@ -96,6 +97,7 @@ leiden <- function(object,
                    resolution_parameter = 1,
                    seed = NULL,
                    n_iterations = 2L,
+                   max_comm_size = 0L,
                    degree_as_node_size = FALSE,
                    laplacian = FALSE) {
     UseMethod("leiden", object)
@@ -121,57 +123,46 @@ leiden.matrix <- function(object,
                           resolution_parameter = 1,
                           seed = NULL,
                           n_iterations = 2L,
+                          max_comm_size = 0L,
                           degree_as_node_size = FALSE,
                           laplacian = FALSE
 ) {
+    if(length(partition_type) > 1) partition_type <- partition_type[[1]][1]
+    partition_type <- match.arg(partition_type)
+
     #import python modules with reticulate
-    numpy <- reticulate::import("numpy", delay_load = TRUE)
+    numpy <- import("numpy", delay_load = TRUE)
     leidenalg <- import("leidenalg", delay_load = TRUE)
     ig <- import("igraph", delay_load = TRUE)
 
     #convert matrix input (corrects for sparse matrix input)
     if(is.matrix(object) || is(object, "dgCMatrix")){
-        adj_mat <- object
+        object <- object
     } else{
-        adj_mat <- as.matrix(object)
+        object <- as.matrix(object)
     }
 
     #compute weights if non-binary adjacency matrix given
-    is_pure_adj <- all(as.logical(adj_mat) == adj_mat)
+    is_pure_adj <- all(as.logical(unlist(object)) == object)
     if (is.null(weights) && !is_pure_adj) {
-        if(!is.matrix(object)) adj_mat <- as.matrix(adj_mat)
+        if(!is.matrix(object)) object <- as.matrix(object)
         #assign weights to edges (without dependancy on igraph)
-        t_mat <- t(adj_mat)
+        t_mat <- t(object)
         weights <- t_mat[t_mat!=0]
         #remove zeroes from rows of matrix and return vector of length edges
     }
 
-    ##convert to python numpy.ndarray, then a list
-    adj_mat_py <- r_to_py(adj_mat, convert = T)
-    if(is(object, "dgCMatrix")){
-        adj_mat_py <- adj_mat_py$toarray()
-    }
-    adj_mat_py <- adj_mat_py$tolist()
-
-    #convert graph structure to a Python compatible object
-    GraphClass <- if (!is.null(weights) && !is_pure_adj){
-        ig$Graph$Weighted_Adjacency
-    } else {
-        ig$Graph$Adjacency
-    }
-    snn_graph <- GraphClass(adj_mat_py)
-    # test performance for computing matrix to graph in R
-    # other option is to passing snn_graph to Python
+    py_graph <- make_py_graph(object, weights = weights)
 
     #compute partitions
-    #compute partitions
-    partition <- find_partition(snn_graph, partition_type = partition_type,
-                                initial_membership = initial_membership ,
+    partition <- find_partition(py_graph, partition_type = partition_type,
+                                initial_membership = initial_membership,
                                 weights = weights,
                                 node_sizes = node_sizes,
                                 resolution_parameter = resolution_parameter,
                                 seed = seed,
                                 n_iterations = n_iterations,
+                                max_comm_size = max_comm_size,
                                 degree_as_node_size = degree_as_node_size
     )
     partition
@@ -202,17 +193,18 @@ leiden.Matrix <- function(object,
                           resolution_parameter = 1,
                           seed = NULL,
                           n_iterations = 2L,
+                          max_comm_size = 0L,
                           degree_as_node_size = FALSE,
                           laplacian = FALSE
 ) {
     #cast to sparse matrix
-    adj_mat <- as(object, "dgCMatrix")
+    object <- as(object, "dgCMatrix")
     #run as igraph object (passes to reticulate)
     if(is.null(weights)){
-        object <- graph_from_adjacency_matrix(adjmatrix = adj_mat, weighted = TRUE)
+        object <- graph_from_adjacency_matrix(adjmatrix = object, weighted = TRUE)
         weights <- edge_attr(object)$weight
     } else {
-        object <- graph_from_adjacency_matrix(adjmatrix = adj_mat, weighted = TRUE)
+        object <- graph_from_adjacency_matrix(adjmatrix = object, weighted = TRUE)
         object <- set_edge_attr(object, "weight", index=E(object), weights)
     }
 
@@ -228,10 +220,94 @@ leiden.Matrix <- function(object,
     )
 }
 
+##' @importFrom igraph graph_from_adjacency_matrix edge_attr set_edge_attr E is.igraph
+##' @importFrom methods as
+##' @importClassesFrom Matrix dgCMatrix dgeMatrix
+##' @export
+leiden.list <- function(object,
+                          partition_type = c(
+                              'RBConfigurationVertexPartition',
+                              'ModularityVertexPartition',
+                              'RBERVertexPartition',
+                              'CPMVertexPartition',
+                              'MutableVertexPartition',
+                              'SignificanceVertexPartition',
+                              'SurpriseVertexPartition',
+                              'ModularityVertexPartition.Bipartite',
+                              'CPMVertexPartition.Bipartite'
+                          ),
+                          initial_membership = NULL,
+                          weights = NULL,
+                          node_sizes = NULL,
+                          resolution_parameter = 1,
+                          seed = NULL,
+                          n_iterations = 2L,
+                          max_comm_size = 0L,
+                          degree_as_node_size = FALSE,
+                          laplacian = FALSE
+) {
+    if(length(partition_type) > 1) partition_type <- partition_type[[1]][1]
+    partition_type <- match.arg(partition_type)
+
+    if(length(object) == 1 || is.igraph(object)){
+        if(!is.igraph(object) && is.list(object)) object <- object[[1]]
+        partition <- leiden.igraph(object,
+                            partition_type = partition_type,
+                            weights = weights,
+                            node_sizes = node_sizes,
+                            resolution_parameter = resolution_parameter,
+                            seed = seed,
+                            n_iterations = n_iterations,
+                            max_comm_size = max_comm_size,
+                            degree_as_node_size = degree_as_node_size,
+                            laplacian = laplacian
+        )
+    } else{
+
+        #import python modules with reticulate
+        numpy <- reticulate::import("numpy", delay_load = TRUE)
+        leidenalg <- import("leidenalg", delay_load = TRUE)
+        ig <- import("igraph", delay_load = TRUE)
+
+        object <- lapply(object, function(graph){
+            if(is.matrix(graph) || is(graph, "dgCMatrix")){
+                graph_from_adjacency_matrix(graph)
+            } else {
+                graph
+            }
+        })
+
+        names(object) <- c()
+
+        py_list <- r_to_py(lapply(object, function(r_graph){
+            make_py_graph(r_graph, weights = weights)
+        }))
+
+
+        if(partition_type == 'ModularityVertexPartition.Bipartite') partition_type <- "ModularityVertexPartition"
+        if(partition_type == 'CPMVertexPartition.Bipartite') partition_type <- "CPMVertexPartition"
+        partition_type <- gsub(".Bipartite", "", partition_type)
+        partition_type <- gsub(".Multiplex", "", partition_type)
+
+        #compute partitions with reticulate
+        partition <- find_partition_multiplex(py_list, partition_type = partition_type,
+                                    initial_membership = initial_membership,
+                                    weights = weights,
+                                    node_sizes = node_sizes,
+                                    resolution_parameter = resolution_parameter,
+                                    seed = seed,
+                                    n_iterations = n_iterations,
+                                    max_comm_size = max_comm_size,
+                                    degree_as_node_size = degree_as_node_size
+        )
+    }
+    partition
+}
+
 ##' @export
 leiden.default <- leiden.matrix
 
-##' @importFrom igraph V as_edgelist is.weighted is.named edge_attr as_adjacency_matrix laplacian_matrix vertex_attr is_bipartite bipartite_mapping set_vertex_attr simplify
+##' @importFrom igraph V as_edgelist is.weighted is.named edge_attr as_adjacency_matrix laplacian_matrix vertex_attr is_bipartite bipartite_mapping set_vertex_attr simplify is_named is_weighted
 ##' @export
 leiden.igraph <- function(object,
                           partition_type = c(
@@ -251,6 +327,7 @@ leiden.igraph <- function(object,
                           resolution_parameter = 1,
                           seed = NULL,
                           n_iterations = 2L,
+                          max_comm_size = 0L,
                           degree_as_node_size = FALSE,
                           laplacian = FALSE
 ) {
@@ -258,6 +335,10 @@ leiden.igraph <- function(object,
     numpy <- reticulate::import("numpy", delay_load = TRUE)
     leidenalg <- import("leidenalg", delay_load = TRUE)
     ig <- import("igraph", delay_load = TRUE)
+
+    #default partition
+    if(length(partition_type) > 1) partition_type <- partition_type[[1]][1]
+    partition_type <- match.arg(partition_type)
 
     ##convert to python numpy.ndarray, then a list
     if(!is.named(object)){
@@ -273,10 +354,6 @@ leiden.igraph <- function(object,
         edgelist[[ii]] <- as.character(edges[ii,])
     }
 
-    snn_graph <- ig$Graph()
-    snn_graph$add_vertices(r_to_py(vertices))
-    snn_graph$add_edges(r_to_py(edgelist))
-
     #derive Laplacian
     if(laplacian == TRUE){
         object <- simplify(object, remove.multiple = TRUE, remove.loops = TRUE)
@@ -287,12 +364,7 @@ leiden.igraph <- function(object,
         }
     }
 
-    #compute weights if weighted graph given
-    if(is.weighted(object)){
-        #assign weights to edges (without dependancy on igraph)
-        weights <- r_to_py(edge_attr(object)$weight)
-        snn_graph$es$set_attribute_values('weight', weights)
-    }
+    py_graph <- make_py_graph(object, weights = weights)
 
     if(length(partition_type) > 1) partition_type <- partition_type[1]
     if(partition_type == "ModularityVertexPartition.Bipartite"){
@@ -319,21 +391,19 @@ leiden.igraph <- function(object,
     }
 
     if(!is.null(vertex_attr(object, "type")) || is_bipartite(object)){
-        type <- as.integer(V(object)$type)
-        snn_graph$vs$set_attribute_values('type', r_to_py(as.integer(type)))
+        type <- as.integer(unlist(V(object)$type))
+        py_graph$vs$set_attribute_values('type', r_to_py(as.integer(type)))
     }
 
-    # from here is the same as method for matrix
-    # would be better to refactor to call from matrix methof
-
     #compute partitions
-    partition <- find_partition(snn_graph, partition_type = partition_type,
+    partition <- find_partition(py_graph, partition_type = partition_type,
                                 initial_membership = initial_membership ,
                                 weights = weights,
                                 node_sizes = node_sizes,
                                 resolution_parameter = resolution_parameter,
                                 seed = seed,
                                 n_iterations = n_iterations,
+                                max_comm_size = max_comm_size,
                                 degree_as_node_size = degree_as_node_size
     )
     partition
@@ -378,6 +448,7 @@ numpy <- NULL
                         install.packages("devtools",  quiet = TRUE)
                         devtools::install_github("rstudio/reticulate", ref = "86ebb56",  quiet = TRUE)
                         if(!reticulate::py_module_available("numpy")) suppressWarnings(suppressMessages(reticulate::conda_install(envname = "r-reticulate", packages = "numpy")))
+                        if(!reticulate::py_module_available("pandas")) suppressWarnings(suppressMessages(reticulate::conda_install(envname = "r-reticulate", packages = "pandas")))
                         if(!reticulate::py_module_available("igraph")) suppressWarnings(suppressMessages(reticulate::conda_install(envname = "r-reticulate", packages = "python-igraph")))
                         if(!reticulate::py_module_available("mkl")) suppressWarnings(suppressMessages(reticulate::conda_install(envname = "r-reticulate", packages = "mkl", channel = "intel")))
                         if(!reticulate::py_module_available("umap")) suppressWarnings(suppressMessages(reticulate::conda_install(envname = "r-reticulate", packages = "umap-learn", channel = "conda-forge")))
@@ -387,6 +458,7 @@ numpy <- NULL
                         utils::install.packages("reticulate",  quiet = TRUE)
                     } else {
                         if(!reticulate::py_module_available("numpy")) suppressWarnings(suppressMessages(reticulate::conda_install("r-reticulate", "numpy")))
+                        if(!reticulate::py_module_available("pandas")) suppressWarnings(suppressMessages(reticulate::conda_install("r-reticulate", "pandas")))
                         if(!reticulate::py_module_available("igraph")) suppressWarnings(suppressMessages(reticulate::conda_install("r-reticulate", "python-igraph")))
                         if(!reticulate::py_module_available("umap")) suppressWarnings(suppressMessages(reticulate::conda_install("r-reticulate", "umap-learn", forge = TRUE)))
                         if(!reticulate::py_module_available("leidenalg")) suppressWarnings(suppressMessages(reticulate::conda_install("r-reticulate", "leidenalg", forge = TRUE)))
@@ -405,6 +477,7 @@ numpy <- NULL
                     # system("conda init")
                     # system("conda activate r-reticulate")
                     if(!reticulate::py_module_available("numpy")) suppressWarnings(suppressMessages(reticulate::py_install("numpy")))
+                    if(!reticulate::py_module_available("pandas")) suppressWarnings(suppressMessages(reticulate::py_install("pandas")))
                     if(!reticulate::py_module_available("igraph")) suppressWarnings(suppressMessages(reticulate::py_install("python-igraph", method = method, conda = conda)))
                     if(!reticulate::py_module_available("umap")) suppressWarnings(suppressMessages(reticulate::py_install("umap-learn")))
                     if(!reticulate::py_module_available("leidenalg")) suppressWarnings(suppressMessages(reticulate::py_install("leidenalg", method = method, conda = conda, forge = TRUE)))
@@ -441,6 +514,7 @@ numpy <- NULL
         if (modules) {
             ## assignment in parent environment!
             numpy <- reticulate::import("numpy", delay_load = TRUE)
+            pd <- reticulate::import("pandas", delay_load = TRUE)
             leidenalg <- reticulate::import("leidenalg", delay_load = TRUE)
             ig <- reticulate::import("igraph", delay_load = TRUE)
         }