+++ NEWS.md
@@ -1,3 +1,9 @@
+# leiden 0.3.6
+
+* add methods for multiplex community detection from a list of graphs
+
+* add support for maximum community size (depends on leidenalg 0.8.2), if available
+
 # leiden 0.3.5
 
 * background changes to build vignettes on CRAN